(chore) Consistently export component property types.
