+++ containers/mosaic/src/components/Drawer/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Drawer";
+export * from "./DrawerTypes";
