+++ containers/mosaic/src/components/DrawerContent/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./DrawerContent";
+export * from "./DrawerTypes";
