+++ containers/mosaic/src/components/Menu/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Menu";
+export * from "./MenuTypes";
