+++ containers/mosaic/src/components/PageHeader/index.ts
@@ -1,3 +1,3 @@
 export { default } from "./PageHeader";
 export * from "./PageHeader";
-export * from "./PageHeader";
+export * from "./PageHeaderTypes";
