+++ containers/mosaic/src/components/Spinner/SpinnerTypes.tsx
@@ -1,0 +1,4 @@
+export interface SpinnerProps {
+	className?: string;
+	progress?: number;
+}
