+++ containers/mosaic/src/components/Spinner/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Spinner";
+export * from "./SpinnerTypes";
